Moved yara updates to receiver module
