+++ bin/Minerva/receiver/__init__.py
@@ -1,2 +1,3 @@
 from .listener import EventReceiver, EventPublisher
 from .worker import EventWorker
+from .watchlist import Watchlist
